去除 caffe 对 boost 的依赖. 增加 opencv2.

+++ include/caffe/util/mkl_alternate.hpp
@@ -49,7 +49,7 @@ DEFINE_VSL_UNARY_FUNC(Abs, y[i] = fabs(a[i]));
     v##name<float>(n, a, b, y); \
   } \
   inline void vd##name( \
-      const int n, const double* a, const float b, double* y) { \
+      const int n, const double* a, const double b, double* y) { \
     v##name<double>(n, a, b, y); \
   }
